+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # C# Device SDK
-- 
+- Add new UploadToBlobAsync API implementing upload of a stream to Azure Storage.
 # General
 -